I2C Master Read/Write functionality
- Added Aardvark I2C discovery board commands in user_app_1.

+++ stm32f407vgt6/drivers/rcc_driver.h
@@ -48,7 +48,7 @@ ClockSourceEnum RCC_GetSystemClockSource();
 //--------------------------------------------------------------
 // Get the frequency of a specific clock type
 //--------------------------------------------------------------
-ULONG RCC_GetClockFrequency(
+UINT RCC_GetClockFrequency(
    ClockTypeEnum eClockType_);
 
 #endif // RCC_DRIVER_H_
